# OzonApi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AviaSales API",
    "description": "REST API for AviaSales booking system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/tickets/": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Create Ticket",
        "operationId": "create_ticket_api_tickets__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "Read Tickets",
        "operationId": "read_tickets_api_tickets__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  },
                  "title": "Response Read Tickets Api Tickets  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "Read Ticket",
        "operationId": "read_ticket_api_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tickets"
        ],
        "summary": "Update Ticket",
        "operationId": "update_ticket_api_tickets__ticket_id__put",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tickets"
        ],
        "summary": "Delete Ticket",
        "operationId": "delete_ticket_api_tickets__ticket_id__delete",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Ticket": {
        "properties": {
          "booking_id": {
            "type": "string",
            "format": "uuid",
            "title": "Booking Id"
          },
          "passenger_id": {
            "type": "string",
            "format": "uuid",
            "title": "Passenger Id"
          },
          "fare_id": {
            "type": "string",
            "format": "uuid",
            "title": "Fare Id"
          },
          "passenger_document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Passenger Document Id"
          },
          "seat_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seat Number"
          },
          "ticket_number": {
            "type": "string",
            "title": "Ticket Number"
          },
          "ticket_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ticket Id"
          },
          "created_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Created Datetime"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "passenger_id",
          "fare_id",
          "passenger_document_id",
          "ticket_number",
          "ticket_id",
          "created_datetime"
        ],
        "title": "Ticket"
      },
      "TicketCreate": {
        "properties": {
          "booking_id": {
            "type": "string",
            "format": "uuid",
            "title": "Booking Id"
          },
          "passenger_id": {
            "type": "string",
            "format": "uuid",
            "title": "Passenger Id"
          },
          "fare_id": {
            "type": "string",
            "format": "uuid",
            "title": "Fare Id"
          },
          "passenger_document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Passenger Document Id"
          },
          "seat_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seat Number"
          },
          "ticket_number": {
            "type": "string",
            "title": "Ticket Number"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "passenger_id",
          "fare_id",
          "passenger_document_id",
          "ticket_number"
        ],
        "title": "TicketCreate"
      },
      "TicketUpdate": {
        "properties": {
          "seat_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seat Number"
          },
          "ticket_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Number"
          }
        },
        "type": "object",
        "title": "TicketUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}